requirements.txt
Flask==2.0.3
hazelcast-python-client=5.1
